Remove questions table in migrations/seeds, improve migration
"Up" migrations will no longer drop existing table. Previously this
meant deploying a change would wipe the old database.

Also, drop reliance on unnecessary `is_playing` field when migrating.
This will have to be changed in the server code.

Drop table creation/seeding of the questions table as we are using
Google Sheets as the sole source of truth.

Signed-off-by: Christina Wang <christina@balena.io>