Change addOrUpdatePlayer to addPlayer
Since we removed leave/join functionality in favor of
having players handle their participation via Flowdock,
we don't need to update player data, only add new handles.

The leave/join functionality was closely tied with the
old is_playing field in the players table, and this was
removed thus we're removing references to it in addPlayer.

Signed-off-by: Christina Wang <christina@balena.io>